requirements.txt
fastapi
uvicorn
livekit
livekit-api
numpy
soundfile
python-dotenv
